Added a supplemental "pac" module.
Right now the public efm32hg-pac crate is not built on the latest SVD
file from the chip maker. The newest version includes the NVIC_PRIO_BITS
value but the efm32hg-pac crate is hard to build due to using unstable
tools.

This module just patches the existing module with the necessary value on
top of the re-exported stuff from the crate.