gpio: Added all routed Tomu pins.
Also, used the paste macro to reduce the amount of input data necessary
to generate the code. Only the bank (a, b, c, ...), index (0..15), and
half (low, high) are needed for now. Hopefully there's a simple way to
get rid of the half part since it's basically an expression of the
index.

Also, corrected the default state for all pins which is Input<Disabled>
and added mode switchers to Input<Disabled> and Output<OpenDrain>.